requirements.txt
biopython
tensorflow>=2.11
numpy
pandas
scikit-learn
matplotlib
seaborn
imageio
tqdm
